refactor: rename medicalHistory module to medical_history and implement new medical record models
